depends: refactor bash array parsing.
use strtok instead of strsep because it will consider multiple
successive delimeters to be a single delimeter.

use strcspn to find the length of the string containing valid characters
up to the first found reject character. Using this length, replace the first
reject character with a '\0'. this replaces the kludgy for loop that iterated
over each character, calling strchr on each iteration.