depends: unify dependency parsing functions
declare an enum of extended package attributes. when extended attributes
are requested, create an alpm_list_t**[6] and use enums to point to the
destination alpm_list_t where the parsed dependencies will be added. if
a list is not specified for one of the array elements, parsing for that
attribute is skipped.