completely overhaul output methodology.
Instead of using an if/then every time we go to print, use pacman-esqe
wrapper functions for the printf family. We also allocate strings which
describe our output and may or may not be of length > 0, but will always
be initialized. This removes the kludgy cprintf functions which were
inflexible and not extensible.