refactor per thread function selection
create struct task_t to describe the process. this defaults to no output
and thread_query for processing, but is overridden on a per operation
basis. This saves parsing the opmask for every argument and knowing
ahead of time what needs to be done.

similarly, on the output side, the presence of a task.printfn means we
need to print results in the results list and describes exactly how to
do it. this is passed on to refactored print_results and print_pkg_*
functions. print_results does the looping, and print_pkg_* functions now
only print the contents of a single package, as their name implies.