return zero for success on all operations except update
This introduces a somewhat nasty hack, but it shouldn't be an issue
given the scope of cower. Famous last words. We define success as:

- a query or download returns something
- update (without --download) returns nothing

Because of this opposing behavior, we have to flip the result of a pure
update request, as the behavior of our thread model returns nothing when
an update (correctly) finds no updates on the AUR.

If only redefining success was always this easy.