only respect the official binary repos
This fixes issues where people having third party binary repos with
binary AUR packages are unable to download the backing AUR taurballs,
receiving the "erroneous" warning: ":: foo is available in bar".

This does introduce hard coding of repo names, but this is a lower
overhead solution than my other idea of checking both the AUR and ALPM
before deciding whether or not to download a package. There will never
be a package by the same name in both the AUR and the official repos.