don't assume an array element ends with a quote
This fixes issues with PKGBUILDs which have escaped newlines in
arrays when the newline escape is adjacent to the line-ending array
element. Examples: ripmake, recorditnow, xastir, uzbl-git.

NOTE: This assumes that we never encounter an unquoted array element
that ends with a backslash. grep'ing through the AUR as of this commit
shows that this never occurs.