Partial revert "only respect the official binary repos"
This reverts commit a74096056d66b5a2260e508542267f2ae840b61a.

This is what happens when you don't test thoroughly. alpm registers a
repo that doesn't exist, and libarchive leaks like crazy. Not a huge
deal, but we'd like to be somewhat clean.

New solution: meet somewhere in the middle. Maintain a list of official
repos, and validate the repos parsed out of pacman.conf against this
list.