lock call to alpm_find_satisfier in resolve_dependencies
use a separate lock from the earlier lock in the same function, since
they're not hitting the same data structures. This isn't a frequently
hit issue, but it does occasionally result in an alpm segfault or double
free, so it's worth fixing. test case:

  -dd wayland-git libgl-git

The key is that libgl-git is a dependency of wayland-git.