remove (almost) all alpm locks
This is possible due to a lot of const qualifiers that Dan has added as
well as a mutex that will be opportunisticly protecting lazy db
population. We leave behind one mutex protecting sync DB access as we
don't want to hackishly lazy load all of them until we need to.